+++ CHANGELOG.md
@@ -34,6 +34,7 @@ your changes, such as:
 
 - [public] [both] [updated] add a new feature
 ## [Unreleased]
+- [public] [both] [doc] add netping plugin
 - [public] [both] [doc] add prometheus fetcher usecase doc
 - [public] [both] [doc] add telegraf receiver usecase doc
 - [public] [both] [updated] enhance memory cost for stdout collection